set up to be good w firebase
