--- src/components/SubAccounts/Individual.js
@@ -1,11 +1,0 @@
-import React from 'react';
-
-export default function Individual(props) {
-    return(
-        <React.Fragment>
-            <Typography variant='h1'>
-                Subaccount Name: 
-            </Typography>
-        </React.Fragment>
-    )
-}
