+++ src/components/SubAccounts/SubAccounts.js
@@ -24,6 +24,8 @@ import InfoIcon from '@material-ui/icons/Info';
 import Progress from '../Dashboard/Progress';
 import PieCharts from '../Dashboard/PieCharts';
 import KeyboardBackspaceIcon from '@material-ui/icons/KeyboardBackspace';
+import { withFirebase } from '../Firebase';
+import IndividualPage from './IndividualPage';
 
 
 const useStyles = makeStyles((theme) => ({
@@ -51,99 +53,99 @@ const useStyles = makeStyles((theme) => ({
     },
 }));
 
-const IndividualPage = ({name, chartInfo, total}) => {
-    return(
-      <div>
-          <Grid container spacing={3} alignItems='center'>
+// const IndividualPage = ({name, chartInfo, total}) => {
+//     return(
+//       <div>
+//           <Grid container spacing={3} alignItems='center'>
 
-              <Grid item xs={12} style={{
-                  marginLeft: 500,
-              }}>
-                <Typography variant='h1'>{name}</Typography>
-              </Grid>
+//               <Grid item xs={12} style={{
+//                   marginLeft: 500,
+//               }}>
+//                 <Typography variant='h1'>{name}</Typography>
+//               </Grid>
 
-              <Grid item xs={12} style={{
-                  marginLeft: 490,
-                  marginTop: 10
-              }}>
-                <Typography variant='h3' style={{
-                    borderStyle: 'solid', 
-                    borderColor: "#88bb00",
-                    borderWidth: 10,
-                    padding: 10,
-                    width: 460
-                }}>Total Subaccount Value: ${total}</Typography>
-              </Grid>
+//               <Grid item xs={12} style={{
+//                   marginLeft: 490,
+//                   marginTop: 10
+//               }}>
+//                 <Typography variant='h3' style={{
+//                     borderStyle: 'solid', 
+//                     borderColor: "#88bb00",
+//                     borderWidth: 10,
+//                     padding: 10,
+//                     width: 460
+//                 }}>Total Subaccount Value: ${total}</Typography>
+//               </Grid>
 
-              <Grid item xs={12} style={{
-                  marginLeft: 610,
-                  marginTop: -10
-              }}>
-                  <Typography variant='h4'>Progress to Goal: $23,500</Typography>
-              </Grid>
+//               <Grid item xs={12} style={{
+//                   marginLeft: 610,
+//                   marginTop: -10
+//               }}>
+//                   <Typography variant='h4'>Progress to Goal: $23,500</Typography>
+//               </Grid>
 
-              <Grid item xs={12} style={{
-                  marginLeft: 470,
-                  marginTop: -10
-              }}>
-                <Progress num="60"></Progress>
-              </Grid>
+//               <Grid item xs={12} style={{
+//                   marginLeft: 470,
+//                   marginTop: -10
+//               }}>
+//                 <Progress num="60"></Progress>
+//               </Grid>
 
-              <Grid item xs={12} style={{
-                  marginTop: 10,
-                  marginLeft: 300
-              }}>
-                <div style={{
-                        height: 800, 
-                        width: 800
-                        }}>
-                <PieCharts data={data1} options={options1}></PieCharts>
-                </div>
-              </Grid>
-          </Grid>
-    </div>
-    )
-  }
+//               <Grid item xs={12} style={{
+//                   marginTop: 10,
+//                   marginLeft: 300
+//               }}>
+//                 <div style={{
+//                         height: 800, 
+//                         width: 800
+//                         }}>
+//                 <PieCharts data={data1} options={options1}></PieCharts>
+//                 </div>
+//               </Grid>
+//           </Grid>
+//     </div>
+//     )
+//   }
 
-    // First Pie Chart data
-    const data1 = {
-        labels: [
-            'GOOG',
-            'SBUX',
-            'EFX'
-        ],
-        datasets: [{
-            data: [30, 50, 20],
-            backgroundColor: [
-            '#47c3d4',
-            '#cc4e00',
-            '#ffc425'
-            ],
-            hoverBackgroundColor: [
-            '#47c3d4',
-            '#cc4e00',
-            '#ffc425'
-            ]
-        }]
-    };
+//     // First Pie Chart data
+//     const data1 = {
+//         labels: [
+//             'GOOG',
+//             'SBUX',
+//             'EFX'
+//         ],
+//         datasets: [{
+//             data: [30, 50, 20],
+//             backgroundColor: [
+//             '#47c3d4',
+//             '#cc4e00',
+//             '#ffc425'
+//             ],
+//             hoverBackgroundColor: [
+//             '#47c3d4',
+//             '#cc4e00',
+//             '#ffc425'
+//             ]
+//         }]
+//     };
 
-  // First Pie Chart Options
-  const options1 = {
-      legend: {
-          position: 'bottom',
-          labels: {
-              fontSize: 20
-          }
-      }, 
-      title: {
-          display: true,
-          position: 'top',
-          fontSize: 24,
-          text: 'Distribution of funds'
-      }
-  }
+//   // First Pie Chart Options
+//   const options1 = {
+//       legend: {
+//           position: 'bottom',
+//           labels: {
+//               fontSize: 20
+//           }
+//       }, 
+//       title: {
+//           display: true,
+//           position: 'top',
+//           fontSize: 24,
+//           text: 'Distribution of funds'
+//       }
+//   }
 
-export default function SubAccounts({page}) {
+  function SubAccounts(props) {
     const classes = useStyles();
     const [view, setView] = useState("Dashboard");
     const [num, setNum] = useState(0);
@@ -204,8 +206,8 @@ export default function SubAccounts({page}) {
                                         <InfoIcon fontSize='medium'></InfoIcon>
                                         </Button>
                                         <Menu {...bindMenu(popupState)}>
-                                        <MenuItem onClick={handleIndividual}>Wedding</MenuItem>
-                                        <MenuItem onClick={handleIndividual}>New Car</MenuItem>
+                                        <MenuItem onClick={handleIndividual}>{authUser.email}</MenuItem>
+                                        <MenuItem onClick={handleIndividual}>{props.firebase.user.name}</MenuItem>
                                         <MenuItem onClick={handleIndividual}>Europe Vacation</MenuItem>
                                         </Menu>
                                     </React.Fragment>
@@ -269,14 +271,19 @@ export default function SubAccounts({page}) {
           }  
           {num == 3 ? (
               <div>
-                    <Button onClick={handleMain} variant="contained" style={{
-                        backgroundColor: "#528400",
-                        color: 'white',  
-                        marginLeft: 20, 
-                        marginTop: 20
-                    }}><KeyboardBackspaceIcon fontSize='medium'></KeyboardBackspaceIcon></Button>
-                  <IndividualPage name='Name from Database' total='20,137'></IndividualPage>
-
+                    <AuthUserContext.Consumer>
+                    {authUser => (
+                        <div>
+                            <Button onClick={handleMain} variant="contained" style={{
+                                backgroundColor: "#528400",
+                                color: 'white',  
+                                marginLeft: 20, 
+                                marginTop: 20
+                            }}><KeyboardBackspaceIcon fontSize='medium'></KeyboardBackspaceIcon></Button>
+                        <IndividualPage name='firebase name' total='20,137'></IndividualPage>
+                        </div>
+                    )}
+                    </AuthUserContext.Consumer>
                   </div>
           ) : (
             <div></div>
@@ -288,6 +295,7 @@ export default function SubAccounts({page}) {
     )
 }
 
+export default withFirebase(SubAccounts);
